Add new configuration functions for Seaglider data loading.
Seaglider data loading for real time and delayed time modes
is configured in new functions:
  - `configRTFileOptionsSeaglider`
  - `configDTFileOptionsSeaglider`

The name pattern of files to load is specified as a regular expression,
and parameters and data columns to load might be chosen.
The path to the files is configured in the same function for all kind of
gliders (`configRTPathsLocal` and `configDTPathsLocal`).