Add FTP-like class for SFTP connections based on libssh.
Class `FTP` provides an interface to operate on files on a remote
server through the FTP protocol. There is no equivalent for SFTP.

Implement the same interface of `FTP` for the SFTP protocol
in a new class `SFTP` using the C library libssh.

Include a helper script to compile the required mex interface to libssh.