Change depth variable selection non-CTD transect plots and data gridding.
The depth measurements derived from the CTD readings might not be aligned
in time with the measurements from sensors other than the CTD
(e.g. chlorophyll and turbidity). Thus plots and processes that
require a valid depth position together with a sensor measurement
might miss the misaligned data.

To show all the recorded data for that sensors, change the default depth
selection to navigation depth, which might be filled by interpolation
hence providing a depth measurement for all sensor readings.

Another option would be to add a `depth_ctd_filling` option
to interpolate the depth measurements derived from the CTD.