Improve cast identification and profile validation.
Modify cast identification algorithm to handle non-smooth depth
inputs, to allow the use of the navigation depth data instead
of the ones derived from the filtered pressure measured by the CTD.
Introduce new options to discard shake or stall periods,
and to control when consecutive cast segments with the same
directions are joined according the duration and depth range
of the interval in-between.

In addition, identify and flag depth inversions in the validation
of a profile, and add an option to return the profile data with
the invalid rows masked by the desired value. This is required
to remove insignificant depth inversions at the beginning or
at the end of a profile introduced by the use of a different
depth sequence for cast identification.

Update the processing function and the default configuration
files to match the new options.