Extend SeaExplorer support to real time processing.
Real time data retrieval is not defined yet for SeaExplorer gliders.
Presumably the glider will send the files to some ground station
accessible through SSH, and hence probably SFTP.
This situation is very similiar to the Slocum and the Seaglider case.
However, the exact details will not be available in the near future.

Adapt the required parts of the real time processing to handle
SeaExplorer data. At the file retrieval stage, emit a warning
but allow the processing to continue with all the files in the
input ascii directory.