Handle defective position data in SeaExplorer preprocessing.
Position data is defective in SeaExplorer data sets.
According to the manufacturer, SeaExplorer gliders can only get
valid GPS fixes in 'transmitting state', but even in that state
they might produce null coordinates for invalid fixes,
or repetitions of a previous fix too far away in time.

Add a good fix position filter for SeaExplorer data to discard
position entries in wrong states, null positions and repeated fixes,
and include it in the default preprocessing configuration.
Also remove the modified version of NMEA to decimal degrees conversion
that discards the null positions, which is not needed any more.