Fix creation of record dimensions in NetCDF save function and improve performance.
Record dimensions might be specified either with the special `length` value 0
or setting the boolean field `unlimited` to true.
Check the boolean field first to create record dimensions in `savenc`
even if their length is positive (as returned by `loadnc`).

Also reorder code so that all variables are defined first,
and then filled with the data values.
This seems to improve performance.